strands/requirements.txt
strands-python>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
psutil>=5.9.0